phase1/requirements.txt
kafka-python==2.0.2
requests==2.31.0
feedparser==6.0.10
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
elasticsearch==8.12.0
psycopg2-binary==2.9.9
pandas==2.1.4
great-expectations==0.18.8
feast==0.36.0
prometheus-client==0.19.0
apache-airflow==2.8.0
websocket-client==1.7.0
